\name{convoi-package}
\alias{convoi-package}
\alias{convoi}
\docType{package}
\title{
  Convex Optimization in R
}
\description{
  The R package `convoi' will provide simple methods and algorithms for
  convex optimization in R.
}
\details{
\tabular{ll}{
Package: \tab convoi\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2012-01-07\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
  Hans W Borchers

  Maintainer: HwB  <hwborchers@googlemail.com>
}
\references{
  Boyd, S., and L. Vandenberghe (2004). Convex Optimization.
  Cambridge University Press, Cambridge, MA.
}
\keyword{ package }
\seealso{
  The R-Forge project ``Optimization and solving packages''.
  \url{https://r-forge.r-project.org/projects/optimizer/}
}
\examples{
# See the functions in this package.
}
